Auto Sort Feature
Assignment rubric states that JComboBox actionListener must be implemented as a Lambda function. I wanted to be able to sort each time an event was added so that the user would not have to manually re-sort each time.

CHANGES:
- sortEvents() method added to EventListPanel
- ^sorts based on the state of the JComboBox
- sortEvents() called in filterDropDown action listener
- sortEvents() called in updateDisplayPanel()

I really just didn't want to go back and change all this if not allowed.